Moved BaseMenu, UIScreen to respective folders in /components
- Updated all references
