+++ source/components/menu/BattleMenuComponent.cpp
